pyinstaller is driving me mad.
there was one bug that was genuinely ours: by moving the extensions
into ext/, but keeping them (in gentle.spec) in Resources/, shelling
out to them failed. fine. i've fixed that by moving the entire ext/
folder to Resources/. as a hack, ffmpeg is in there too, so the
get_binary function adds ext/ if we're frozen.

i have no idea why pyinstaller all of a sudden (using the python.org
python) decided to put the entire gentle/ directory into the bundle.
there's an egregious hack in make_dmg.sh to work around that.